update install script

+++ install_dotfiles.bash
@@ -1,5 +1,11 @@
+#!/usr/bin/env bash
+
+
 for FILE in .bashrc .bash_aliases .gitconfig
 do
-    cp ~/$FILE ~/$FILE.old
-    cp ~/code/dotfiles/$FILE ~/$FILE
+    if [ -f ~/$FILE ]
+    then
+        cp ~/$FILE ~/$FILE.old
+        cp ~/code/dotfiles/$FILE ~/$FILE
+    fi
 done
